API: Issue delete event on a resource under observation
We make use of the fact that iotivity triggers a GET request with the resource
OC_ACTIVE flag turned off to identify the case when the reason for the GET
request is that a resource was unregistered, but an observer is still attached.
We return OC_STACK_NO_RESOURCE to indicate that the resource is no longer
available for observation.

The client side handles this return value by emitting a "delete" event for the
resource in question.